% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parameters.R
\name{evaluate_parameters}
\alias{evaluate_parameters}
\title{Evaluate estimated parameters}
\usage{
evaluate_parameters(est_par, true_par, criterion = "RMSE")
}
\arguments{
\item{est_par}{a vector or matrix of estimated parameters}

\item{true_par}{a vector or matrix of true parameters}

\item{criterion}{a string specifying the used criterion for comparing the
estimated and true parameters}
}
\value{
a numeric vector of length one
}
\description{
Compare estimated parameters to true parameters
}
\details{
The same \code{criterion}s are implemented for \code{evaluate_parameters}
and \code{\link{evaluate_imputed_values}}.
The possible choices are documented in \code{\link{evaluate_imputed_values}}.
}
\examples{
evaluate_parameters(1:4, 2:5, "RMSE")
}
