% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_expected_values.R
\name{impute_expected_values}
\alias{impute_expected_values}
\title{Impute expected values}
\usage{
impute_expected_values(
  ds,
  mu,
  S,
  stochastic = FALSE,
  M = is.na(ds),
  warn_problematic_rows = FALSE
)
}
\arguments{
\item{ds}{a dataset with missing values}

\item{mu}{vector of expected values for the variables}

\item{S}{covariance matrix of the variables}

\item{stochastic}{logical, should residuals be added to the expected values}

\item{M}{missing data indicator matrix}

\item{warn_problematic_rows}{should warnings be given for problematic rows?}
}
\value{
a dataset of the same type as \code{ds}
}
\description{
Impute the missing values with expected values given the observed values and
estimated parameters assuming a multivariate normal distribution
}
\details{
Normally, this function is called by other imputation function and should not
be called directly. The function imputes the missing values assuming a
multivariate normal distribution. This is equivalent to imputing the least
squares estimate of the missing values in some kind of way.

If no values is observed in a row or the matrix a relevant submatrix of the
covariance matrix (\code{S_22}) is not invertible, the missing values are imputed
with (parts of) \code{mu} (plus a residuum, if \code{stochastich = TRUE}). If
\code{warn_problematic_rows = TRUE}, the problematic rows will be listed in a
warning. Otherwise, they will be imputed without any message.
}
