% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_GMC.R
\name{impute_GMC}
\alias{impute_GMC}
\title{Gaussian mixture clustering imputation}
\usage{
impute_GMC(ds, k_max, imp_max_iter = 10L)
}
\arguments{
\item{ds}{A data frame or matrix with missing values.}

\item{k_max}{maximum number of clusters (called `S` by Ouyang et al. (2004))}

\item{imp_max_iter}{maximum number of iterations for `K_estimate()`}
}
\value{
An object of the same class as \code{ds} with imputed missing values.
}
\description{
Impute missing values in a data frame or a matrix using parameters estimated
via Gaussian mixture clustering
}
\details{
This function performs Gaussin mixture clustering (GMC) imputation as
described by Ouyang et al. (2004).
}
\references{
Ouyang, M., Welsh, W. J., Georgopoulos, P. (2004): Gaussian
  Mixture Clustering and Imputation of Microarray Data.
  \emph{Bioinformatics}, 20(6), 917–923
}
