% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_LS_adaptive.R
\name{impute_LS_adaptive}
\alias{impute_LS_adaptive}
\title{LSimpute_adaptive}
\usage{
impute_LS_adaptive(
  ds,
  k = 10,
  eps = 1e-06,
  min_common_obs = 5,
  r_max_min = 100,
  p_mis_sim = 0.05,
  warn_r_max = TRUE
)
}
\arguments{
\item{ds}{a data frame or matrix with missing values}

\item{k}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{eps}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{min_common_obs}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{r_max_min}{minimum number of nearest genes used for imputation. The
default value (100) corresponds to the choice of Bo et al. (2004).}

\item{p_mis_sim}{percentage of known values that are set \code{NA} to estimate the
mixing coefficient \code{p}. The default value (0.05) corresponds to the choice
of Bo et al. (2004).}

\item{warn_r_max}{should a warning be given, if \code{r_max_min} is set too high?}
}
\value{
An object of the same class as \code{ds} with imputed missing values
}
\description{
Perform LSimpute_adaptive as described by Bo et al. (2004)
}
\details{
This function performs LSimpute_adaptive as described by Bo et al. (2004).
The function assumes that the genes are the rows of \code{ds}.

If the dataset is too small or has too many missing values, there are some
fallback systems implemented. First, if \code{ncol(ds) <= min_common_obs}
(normally, this should not the case!), values are imputed through
\code{\link[=impute_LS_array]{impute_LS_array()}}. Second \code{r_max_min} is automatically adjusted, if it is
too big (but, a warning will be given, if \code{warn_r_max = TRUE}). Third, if
there are not enough observed values in row (less than\code{min_common_obs}), the
calculation of the mixing coefficient is not possible, and missing values of
these rows are imputed with the values from \code{impute_LS_array()}.
}
\examples{
impute_LS_adaptive(data.frame(X = 1:11, Y = c(1:10, NA)))
}
\references{
Bo, T. H., Dysvik, B., & Jonassen, I. (2004). LSimpute: accurate
estimation of missing values in microarray data with least squares methods.
Nucleic acids research, 32(3), e34
}
\seealso{
\code{\link[=impute_LS_gene]{impute_LS_gene()}} and \code{\link[=impute_LS_array]{impute_LS_array()}}, which imputation values
are combined
}
