% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_LS_array.R
\name{impute_LS_array}
\alias{impute_LS_array}
\title{LSimpute_array}
\usage{
impute_LS_array(
  ds,
  k = 10,
  eps = 1e-06,
  min_common_obs = 5,
  ds_impute_LS_gene = NULL
)
}
\arguments{
\item{ds}{a data frame or matrix with missing values}

\item{k}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{eps}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{min_common_obs}{directly passed to \code{\link[=impute_LS_gene]{impute_LS_gene()}}}

\item{ds_impute_LS_gene}{result of imputing \code{ds} with \code{ds_impute_LS_gene()},
if this already exists}
}
\value{
An object of the same class as \code{ds} with imputed missing values
}
\description{
Perform LSimpute_array as described by Bo et al. (2004)
}
\details{
This function performs LSimpute_array as described by Bo et al.
(2004). The function assumes that the genes are the rows of \code{ds}.
}
\examples{
impute_LS_array(data.frame(X = 1:11, Y = c(1:10, NA)))
}
\references{
Bo, T. H., Dysvik, B., & Jonassen, I. (2004). LSimpute: accurate
estimation of missing values in microarray data with least squares methods.
Nucleic acids research, 32(3), e34
}
\seealso{
\code{\link[=impute_LS_gene]{impute_LS_gene()}}, which is used for the first imputation of the
missing values (if \code{ds_impute_LS_gene} is \code{NULL})
}
