% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_location_parameter.R
\name{impute_median}
\alias{impute_median}
\title{Median imputation}
\usage{
impute_median(ds, type = "columnwise", ordered_low = FALSE)
}
\arguments{
\item{ds}{a data frame or matrix with missing values}

\item{type}{one of: "columnwise", "rowwise", "total", "Two-Way" or "Winer"
(see details)}

\item{ordered_low}{logical; used for the calculation of ordered factors (for
details see: \link{median.factor})}
}
\value{
An object of the same class as \code{ds} with imputed missing values
}
\description{
Impute an observed median value for every missing value
}
\details{
This function behaves exactly like \code{\link{impute_mean}}.
The only difference is that it imputes a median instead of a mean.
All \code{type}s from \code{\link{impute_mean}} are also implemented for
\code{impute_median}.
They are documented in \code{\link{impute_mean}} and
\code{\link{apply_imputation}}.
The function \code{\link[stats]{median}} is used for the calculation of
the median values for imputation.
}
\examples{
ds <- data.frame(X = 1:20, Y = ordered(LETTERS[1:20]))
ds_miss <- delete_MCAR(ds, 0.2)
ds_imp <- impute_median(ds_miss)
# completely observed columns can be of any type:
ds_miss_char <- cbind(ds_miss, letters[1:20])
ds_imp_char <- impute_median(ds_miss_char)
}
\seealso{
\code{\link{apply_imputation}} the workhorse for this function

\code{\link[stats]{median}}, \code{\link{median.factor}}

Other location parameter imputation functions: 
\code{\link{impute_mean}()},
\code{\link{impute_mode}()}
}
\concept{location parameter imputation functions}
