% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate missing data}
\usage{
simulate_data(
  n,
  mean = rep(0, nrow(sigma)),
  sigma = diag(length(mean)),
  mech_type = "MCAR",
  p = 0.1,
  cols_mis = seq_len(length(mean)),
  n_mis_stochastic = FALSE,
  p_overall = FALSE,
  mvtnorm_method = "eigen",
  ...
)
}
\arguments{
\item{n}{number of observations for the complete data set}

\item{mean}{mean vector for simulation of complete data}

\item{sigma}{covariance matrix for simulation of complete data}

\item{mech_type}{type of missing data mechanism. This can be any "ending" of
a \code{delete_} function name, e.g. "MCAR" or "MAR_1_to_x".}

\item{p}{A numeric vector with length one or equal to length \code{cols_mis};
the probability that a value is missing.}

\item{cols_mis}{A vector of column names or indices of columns in which
missing values will be created.}

\item{n_mis_stochastic}{Logical, should the number of missing values be
stochastic? If \code{n_mis_stochastic = TRUE}, the number of missing values
for a column with missing values \code{cols_mis[i]} is a random variable
with expected value \code{nrow(ds) * p[i]}. If \code{n_mis_stochastic =
FALSE}, the number of missing values will be deterministic. Normally, the
number of missing values for a column with missing values
\code{cols_mis[i]} is \code{round(nrow(ds) * p[i])}. Possible deviations
from this value, if any exists, are documented in Details.}

\item{p_overall}{Only used, if \code{mech_type = MCAR}. See \code{\link[=delete_MCAR]{delete_MCAR()}} for details.}

\item{mvtnorm_method}{\code{method} argument from \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}}

\item{...}{Further arguments for \code{delete_} function}
}
\value{
A list with
\itemize{
\item \code{ds_comp} the complete data set
\item \code{pars_true} the true parameters for the simulation of the data set
\item \code{ds_mis} the data set with missing values
}
}
\description{
This function first creates a complete data set via \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}.
Then it will create missing values via a \code{delete_} function. The \code{delete_}
function is specified with \code{mech_type}.
}
\examples{
simulate_data(100, rep(0, 3))
# create MNAR censored data with 20 \% missing values
simulate_data(100, rep(0, 4), mech_type = "MNAR_censoring", p = 0.2)
}
