% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_EM.R
\name{impute_EM}
\alias{impute_EM}
\title{EM imputation}
\usage{
impute_EM(
  ds,
  stochastic = TRUE,
  maxits = 1000,
  criterion = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{ds}{a data frame or matrix with missing values}

\item{stochastic}{logical; see details}

\item{maxits}{maximum number of iterations for the EM, passed to
\code{\link[norm:em.norm]{norm::em.norm()}}}

\item{criterion}{if maximum relative difference in parameter estimates is
below this threshold, the EM algorithm stops, argument is directly passed
to \code{\link[norm:em.norm]{norm::em.norm()}}}

\item{verbose}{should messages be given for special cases (see details)}
}
\value{
An object of the same class as \code{ds} with imputed missing values
}
\description{
Impute missing values in a data frame or a matrix using parameters estimated
via EM
}
\details{
At first parameters are estimated via \code{\link[norm:em.norm]{norm::em.norm()}}. Then these
parameters are used in a regression like model to impute the missing values.
If \code{stochachstic = FALSE}, the expected values (given the observed values and
the estimated parameters via EM) is imputed for the missing values of an
object. If \code{stochastic = TRUE} residuals from a multivariate normal
distribution are added to these expected values.

If no values is observed for a row or the required part of the covariance
matrix for the calculation of the expected values is not invertible, parts of
the estimated mu (estimated mean of the variables) will be imputed. If
\code{stochastic = TRUE}, residuals will be added to these values. If
\code{verbose = TRUE}, a message will be given for these rows.
}
\examples{
ds_orig <- mvtnorm::rmvnorm(100, rep(0, 7))
ds_mis <- delete_MCAR(ds_orig, p = 0.2)
ds_imp <- impute_EM(ds_mis, stochastic = FALSE)
}
